fix(timex): 提供并发安全的时间源替换
将 Now 收敛为函数，并通过 SetNowProvider 管理可恢复的时间源替换，避免并发测试和乱序恢复破坏全局状态。

补充并发替换、乱序恢复与零值边界回归测试。

+++ lang/timex/duration.go
@@ -181,7 +181,7 @@ func ParseDuration(s string) (time.Duration, error) {
 		}
 
 		negative := matches[1] == "-"
-		limit := uint64(^uint64(0) >> 1)
+		limit := ^uint64(0) >> 1
 		if negative {
 			limit++
 		}
@@ -211,7 +211,7 @@ func ParseDuration(s string) (time.Duration, error) {
 		}
 
 		if negative {
-			if magnitude == uint64(^uint64(0)>>1)+1 {
+			if magnitude == (^uint64(0)>>1)+1 {
 				return time.Duration(-1 << 63), nil
 			}
 			return -time.Duration(magnitude), nil
